+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* Re-introduce futility pruning and make it depth dependent (4.38 +- 3.25)
+
 ## [12.0]
 
 * SPSA tune - 938 iters at 20.0+0.20 (24.36 +- 8.47)